Squashed commit of the following:
commit 41d04976b35dfa380feba3a7ef9721ef9cc23c64
Merge: 4533bc4 c78f84a
Author: Guilhem Lavaux <guilhem.lavaux@iap.fr>
Date:   Fri Sep 10 12:50:05 2021 +0200

    Merge branch 'guilhem/develop' of github.com:lavaux/EuclidMocksPipeline into guilhem/develop

commit 4533bc40b40b89eab7db2fe8feeaeb32a1d16bd7
Author: Guilhem Lavaux <guilhem.lavaux@iap.fr>
Date:   Fri Sep 10 11:20:23 2021 +0200

    Create base path if it does not exist

commit de6e7f55d9911099e2e8305fe8da57675c037db8
Author: Guilhem Lavaux <guilhem.lavaux@iap.fr>
Date:   Fri Sep 10 11:15:57 2021 +0200

    Remove large products from git

commit c003fd2faa9884d5f3d3bb8ab06603944395125d
Author: Guilhem Lavaux <guilhem.lavaux@iap.fr>
Date:   Fri Sep 10 11:15:17 2021 +0200

    Update product

commit 3a81997fd60fc39bd0a86042aad119bb6e6e3f18
Author: Guilhem Lavaux <guilhem.lavaux@iap.fr>
Date:   Fri Sep 10 11:14:57 2021 +0200

    Fixups

+++ src/euclid_obssys/tool/create_fulloctant_footprint.py
@@ -61,7 +61,7 @@ def createFullOctantFootprint(outdir: str = "Products") -> None:
 
     with DefaultCatalogWrite(footprint_fname) as out_file:
 
-        footprint = footprint.astype([("FOOTPRINT_G", long)])
+        footprint = footprint.astype([("FOOTPRINT_G", int)])
         out_file.set_array("footprint", footprint)
         out_file.add_tag("RES", footprint_res)
         out_file.add_tag("MINZ", footprint_zrange[0])
